+++ tests/integration/cases/call_per_element_false/CommonLisp_call.lisp
@@ -1,0 +1,2 @@
+(defun process (&rest args) (declare (ignore args)) nil)
+(process :data 1)